Moved helper functions to new folder
